refactor: create store for search value
