+++ src/features/searchSlice.ts
@@ -1,0 +1,23 @@
+import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+
+interface SearchState {
+  value: string;
+}
+
+const initialState: SearchState = {
+  value: '',
+};
+
+export const searchSlice = createSlice({
+  name: 'searchValue',
+  initialState,
+  reducers: {
+    setSearchValue: (state, action: PayloadAction<string>) => {
+      state.value = action.payload;
+    },
+  },
+});
+
+export const { setSearchValue } = searchSlice.actions;
+
+export default searchSlice.reducer;
