+++ src/store/store.ts
@@ -1,9 +1,14 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { rickAndMortyApi } from '../api/api';
+import searchSlice from '../features/searchSlice';
 
 export const store = configureStore({
   reducer: {
     [rickAndMortyApi.reducerPath]: rickAndMortyApi.reducer,
+    search: searchSlice,
   },
   middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(rickAndMortyApi.middleware),
 });
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
